Refactor command directory structure, and sendOptionsMessage
